Move settings item widget to components
