+++ gui/pages/settings/settings_page.py
@@ -1,20 +1,13 @@
-from PySide6.QtCore import (
-    QFileInfo,
-    QDir,
-    Signal
-)
 from PySide6.QtWidgets import (
     QWidget,
     QLabel,
-    QPushButton,
 )
 
 from ..page import Page
 from gui.model.settings import app_settings
+from gui.components.settings.settings_item_widget import SettingsItemWidget
 from gui.widgets import (
-    HBoxLayout,
     VBoxLayout,
-    StylableWidget
 )
 from gui.style import constants
 
@@ -132,44 +125,3 @@ class SettingsPage(Page):
         layout.addWidget(info_container_widget)
 
         layout.addStretch()
-
-
-class SettingsItemWidget(StylableWidget):
-    """
-    A widget for a single setting. 
-    Includes a name, current value and button to set it.
-    """
-
-    button_clicked = Signal()
-
-    def __init__(self, name: str, value: str, button: bool = True):
-        """
-        Initialize the widget for a single setting.
-        """
-        super().__init__()
-        self._name = name
-        self._value = value
-        self.setObjectName("settings_item_widget")
-        layout = HBoxLayout(self)
-        layout.setContentsMargins(0,0,0,0)
-        layout.setSpacing(0)
-
-        # Label to show the workspace folderpath
-        self.label = QLabel(self)
-        self._update_label(self._value)
-        layout.addWidget(self.label, 1)
-
-        # Button to select a new workspace
-        if button:
-            self.chooser = QPushButton(f"Set {self._name}")
-            self.chooser.clicked.connect(self.button_clicked)
-            layout.addWidget(self.chooser)
-
-        layout.addSpacing(10)   
-        
-    def _update_label(self, value: str) -> None:
-        """
-        Update the label of a setting when the current value changes. 
-        """
-        self._value = value
-        self.label.setText(f"Current {self._name}: '{self._value}'")
